+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## _Under development_
 
+- `RPC::Server::Dispatcher`
+    - Check for Instance status via the bind address, not the external one.
+    - Added more status and debuggig messages
 - `Browser`
     - Fixed issue causing the removal of cookie HttpOnly flags.
 - `Parser`